Refactor TitleBar and Sidebar components
