+++ app/ui/main_window.py
@@ -6,21 +6,22 @@ Defines the main application window, including the custom title bar and sidebar.
 # ── Imports ─────────────────────────────────────────────────────────────────────
 from typing import TYPE_CHECKING, Callable
 
-from PySide6.QtCore import QPoint, QPropertyAnimation, QRect, QSize, Qt, Signal
+from PySide6.QtCore import QPoint, QPropertyAnimation, QRect, Qt, Signal
 from PySide6.QtGui import QGuiApplication
-from PySide6.QtWidgets import (QButtonGroup, QDialog, QHBoxLayout, QLabel,
-                               QSizePolicy, QSpacerItem, QStackedWidget,
-                               QVBoxLayout, QWidget)
+from PySide6.QtWidgets import (
+    QDialog,
+    QHBoxLayout,
+    QStackedWidget,
+    QVBoxLayout,
+    QWidget,
+)
 
-from app.config import APPLICATION_WINDOW, SIDEBAR
-from app.core.utils import DebugLogger
+from app.config import APPLICATION_WINDOW
 from app.style_manager import ThemeController
-from app.ui.animations import Animator, WindowAnimator
-from app.ui.components import CustomGrip, NavButton
-from app.ui.components.image import AvatarLoader
-from app.ui.helpers.ui_helpers import create_fixed_wrapper
-from app.ui.widgets import CTIcon, CTToolButton
-from app.ui.widgets.helpers import ButtonEffects
+from app.ui.animations import WindowAnimator
+from app.ui.components import CustomGrip
+from app.ui.components.navigation.sidebar import Sidebar
+from app.ui.components.navigation.titlebar import TitleBar
 
 if TYPE_CHECKING:
     from app.core.services import NavigationService
@@ -28,11 +29,6 @@ if TYPE_CHECKING:
 
 # ── Constants ───────────────────────────────────────────────────────────────────
 SETTINGS = APPLICATION_WINDOW["SETTINGS"]
-APP_ICONS = APPLICATION_WINDOW["ICONS"]
-SIDEBAR_ICONS = SIDEBAR["ICONS"]
-START = SIDEBAR["SETTINGS"]["EXPANDED_WIDTH"]
-END = SIDEBAR["SETTINGS"]["COLLAPSED_WIDTH"]
-DURATION = SIDEBAR["SETTINGS"]["DURATION"]
 
 # ── Application Window ──────────────────────────────────────────────────────────
 class MainWindow(QDialog):
@@ -239,333 +235,4 @@ class MainWindow(QDialog):
 
         self.setGeometry(new_rect)
 
-# ── Title Bar  ──────────────────────────────────────────────────────────────────
-class TitleBar(QWidget):
-    """A custom title bar for the frameless main window.
-
-    Provides window control buttons (minimize, maximize, close), a sidebar toggle button,
-    the application logo, and title. It also handles window dragging and maximizing on double-click.
-
-    Attributes:
-        old_pos (QPoint | None): Stores the last mouse position during a drag operation.
-        logo (CTIcon): The application logo icon.
-        title (QLabel): Label displaying the application title.
-        btn_ico_toggle_sidebar (CTToolButton): Button to toggle the sidebar.
-        btn_ico_minimize (CTToolButton): Button to minimize the window.
-        btn_ico_maximize (CTToolButton): Button to maximize/restore the window.
-        btn_ico_close (CTToolButton): Button to close the window.
-
-    Signals:
-        sidebar_toggled: Emitted when the sidebar toggle button is clicked.
-        close_clicked: Emitted when the close button is clicked.
-        minimize_clicked: Emitted when the minimize button is clicked.
-        maximize_clicked: Emitted when the maximize/restore button is clicked.
-    """
-
-
-    # ── Signals ─────────────────────────────────────────────────────────────────────
-    sidebar_toggled  = Signal()  # To toggle sidebar visibility
-    close_clicked    = Signal()  # To signal a close event
-    minimize_clicked = Signal()  # To signal a minimize event
-    maximize_clicked = Signal()  # To signal a maximize/restore event
-
-    def __init__(self, parent):
-        """Initializes the TitleBar.
-
-        Args:
-            parent (QWidget): The parent widget.
-        """
-        # ── Properties ──
-        super().__init__(parent)
-        self.setObjectName("TitleBar")
-        self.setAttribute(Qt.WA_StyledBackground)
-        self.setFixedHeight(60)
-
-        self.old_pos = None  # initialize old position for dragging
-
-        # ── Title Label ──
-        self.logo = CTIcon(
-            file_path = APP_ICONS["LOGO"]["PATH"],
-            icon_size = APP_ICONS["LOGO"]["SIZE"],
-            variant   = APP_ICONS["LOGO"]["STATIC"],
-        )
-        self.logo.setFixedSize(32,32)
-        self.logo.setObjectName("AppLogo")
-        self.title = QLabel(SETTINGS["APP_NAME"])
-
-        # ── Sidebar Toggle Button ──
-        self.btn_ico_toggle_sidebar = CTToolButton(
-            file_path = APP_ICONS["TOGGLE_SIDEBAR"]["PATH"],
-            icon_size = APP_ICONS["TOGGLE_SIDEBAR"]["SIZE"],
-            variant   = APP_ICONS["TOGGLE_SIDEBAR"]["DYNAMIC"],
-            checkable = True,
-        )
-        self.btn_ico_toggle_sidebar.setFixedSize(SETTINGS["BTN_SIZE"])
-        self.btn_ico_toggle_sidebar.setObjectName("BtnToggleSidebar")
-
-        # ── Minimize Button ──
-        self.btn_ico_minimize = CTToolButton(
-            file_path = APP_ICONS["MINIMIZE"]["PATH"],
-            icon_size = APP_ICONS["MINIMIZE"]["SIZE"],
-            variant   = APP_ICONS["MINIMIZE"]["DYNAMIC"],
-        )
-        self.btn_ico_minimize.setFixedSize(SETTINGS["BTN_SIZE"])
-
-        # ── Maximize/Restore Button ──
-        self.btn_ico_maximize = CTToolButton(
-            file_path = APP_ICONS["MAXIMIZE"]["PATH"],
-            icon_size = APP_ICONS["MAXIMIZE"]["SIZE"],
-            variant   = APP_ICONS["MAXIMIZE"]["DYNAMIC"],
-        )
-        self.btn_ico_maximize.setFixedSize(SETTINGS["BTN_SIZE"])
-        
-        # ── Close Button ──
-        self.btn_ico_close = CTToolButton(
-            file_path = APP_ICONS["CLOSE"]["PATH"],
-            icon_size = APP_ICONS["CLOSE"]["SIZE"],
-            variant   = APP_ICONS["CLOSE"]["DYNAMIC"],
-        )
-        self.btn_ico_close.setFixedSize(SETTINGS["BTN_SIZE"])
-        self.btn_ico_close.setObjectName("BtnClose")
-
-        self._build_layout()  # build the layout for the title bar
-        self._connect_signals() # connect button signals
-
-    def _connect_signals(self):
-        """Connects signals from buttons to the TitleBar's signals."""
-        self.btn_ico_toggle_sidebar.clicked.connect(self.sidebar_toggled.emit)
-        self.btn_ico_minimize.clicked.connect(self.minimize_clicked)
-        self.btn_ico_maximize.clicked.connect(self.maximize_clicked)
-        self.btn_ico_close.clicked.connect(self.close_clicked)
-
-    def _build_layout(self):
-        """Builds the layout for the title bar, arranging buttons and title."""
-        title_bar_layout = QHBoxLayout(self)
-        title_bar_layout.addWidget(self.btn_ico_toggle_sidebar)
-        title_bar_layout.addWidget(self.logo)
-        title_bar_layout.addWidget(self.title)
-        title_bar_layout.addStretch(1)  # push buttons to the right
-        title_bar_layout.addWidget(self.btn_ico_minimize)
-        title_bar_layout.addWidget(self.btn_ico_maximize)
-        title_bar_layout.addWidget(self.btn_ico_close)
-        title_bar_layout.setContentsMargins(0, 0, 0, 0) 
-
-    @property
-    def buttons(self):
-        """Returns a dictionary of the control buttons.
-
-        Returns:
-            dict: A dictionary mapping button names to their respective
-                  CTToolButton instances.
-        """
-        return {
-            "toggle_sidebar": self.btn_ico_toggle_sidebar,
-            "minimize": self.btn_ico_minimize,
-            "maximize": self.btn_ico_maximize,
-            "close": self.btn_ico_close,
-        }
-
-    def update_maximize_icon(self, maximized: bool):
-        """Updates the maximize/restore button icon based on window state.
-
-        Args:
-            maximized: True if the window is maximized, False otherwise.
-        """
-        target_path = (
-            APP_ICONS["RESTORE"]["PATH"] if maximized else APP_ICONS["MAXIMIZE"]["PATH"]
-        )
-
-        # reapply hover effects with the correct base icon
-        ButtonEffects.recolor(
-            self.btn_ico_maximize,
-            file_name=target_path,
-            size=self.btn_ico_maximize.iconSize(),
-            variant=SETTINGS["BTN_STYLE"]["DYNAMIC"],
-        )
-
-    def mousePressEvent(self, event):
-        """Handles mouse press events to initiate window dragging.
-
-        Args:
-            event (QMouseEvent): The mouse press event.
-        """
-        # Allow dragging only when the window is not maximized.
-        if event.button() == Qt.LeftButton and not self.window()._is_maximized:
-            self.old_pos = event.globalPos()
-        super().mousePressEvent(event)
-
-    def mouseMoveEvent(self, event):
-        """Handles mouse move events to drag the window.
-
-        Args:
-            event (QMouseEvent): The mouse move event.
-        """
-        if self.old_pos:
-            delta = event.globalPos() - self.old_pos
-            win = self.window()
-            win.move(win.x() + delta.x(), win.y() + delta.y())
-            self.old_pos = event.globalPos()
-        super().mouseMoveEvent(event)
-
-    def mouseReleaseEvent(self, event):
-        """Handles mouse release events to stop dragging.
-
-        Args:
-            event (QMouseEvent): The mouse release event.
-        """
-        self.old_pos = None
-        self.setCursor(Qt.ArrowCursor)
-        super().mouseReleaseEvent(event)
-
-    def mouseDoubleClickEvent(self, event):
-        """Handles a double-click event on the title bar.
-
-        This triggers the maximize/restore functionality.
-
-        Args:
-            event (QMouseEvent): The mouse double-click event.
-        """
-        if event.button() == Qt.LeftButton:
-            self.maximize_clicked.emit()
-        super().mouseDoubleClickEvent(event)
-
-# ── Sidebar ─────────────────────────────────────────────────────────────────────
-class Sidebar(QWidget):
-    """A collapsible sidebar widget for navigation.
-
-    Provides the user avatar, username, and a vertical stack of navigation buttons
-    (e.g., Dashboard, Meal Planner, Settings). It can be expanded or collapsed.
-
-    Attributes:
-        nav_button_group (QButtonGroup): Ensures that only one navigation button is active at a time.
-        _is_expanded (bool): Tracks the current state (expanded or collapsed).
-    """
-
-    def __init__(self, parent=None):
-        """Initializes the Sidebar.
-
-        Args:
-            parent (QWidget, optional): parent widget. Defaults to None.
-        """
-        super().__init__(parent)
-        # properties
-        self.setObjectName("Sidebar")
-        self.setAttribute(Qt.WA_StyledBackground, True)
-        self.setMinimumWidth(0)
-        self.setMaximumWidth(START)  # start expanded
-        self._is_expanded = True
-
-        # main layout setup
-        self.mainLayout = QVBoxLayout(self)
-        self.mainLayout.setContentsMargins(0, 0, 0, 0)
-        self.mainLayout.setSpacing(0)
-
-        # avatar and username
-        self._create_avatar()
-
-        # navigation buttons
-        self.nav_button_group = QButtonGroup(self) # button group for exclusive selection
-        self.nav_button_group.setExclusive(True)
-        self.btn_dashboard = self._create_nav_button("Dashboard")
-        self.btn_meal_planner = self._create_nav_button("Meal Planner")
-        self.btn_view_recipes = self._create_nav_button("View Recipes")
-        self.btn_shopping_list = self._create_nav_button("Shopping List")
-        self.btn_add_recipes = self._create_nav_button("Add Recipes")
-
-        # bottom spacer to push settings/exit to the bottom
-        self.mainLayout.addItem(QSpacerItem(212, 40, QSizePolicy.Minimum, QSizePolicy.Expanding))
-
-        # settings and exit buttons
-        self.btn_settings = self._create_nav_button("Settings", is_checkable=False)
-        self.btn_exit = self._create_nav_button("Exit", is_checkable=False)
-
-    @property
-    def buttons(self) -> dict[str, NavButton]:
-        """Provides access to all buttons in the sidebar.
-
-        Returns:
-            dict[str, NavButton]: A mapping of button object names to their NavButton instances.
-        """
-        return {
-            "btn_dashboard": self.btn_dashboard,
-            "btn_meal_planner": self.btn_meal_planner,
-            "btn_view_recipes": self.btn_view_recipes,
-            "btn_shopping_list": self.btn_shopping_list,
-            "btn_add_recipes": self.btn_add_recipes,
-            "btn_settings": self.btn_settings,
-            "btn_exit": self.btn_exit,
-        }
-    
-    @property
-    def is_expanded(self) -> bool:
-        """Check if the sidebar is currently expanded.
-
-        Returns:
-            bool: True if expanded, False if collapsed.
-        """
-        return self._is_expanded
-    
-    def toggle(self):
-        """Toggle the sidebar's expanded/collapsed state."""
-        DebugLogger.log(f"Toggling sidebar: currently {'expanded' if self._is_expanded else 'collapsed'}")
-
-        start = self.maximumWidth()
-        end = END if self._is_expanded else START
-        duration = DURATION
-
-        Animator.animate_width(self, start, end, duration)
-
-        self._is_expanded = not self._is_expanded
-        
-    # ── Private Methods ─────────────────────────────────────────────────────────────
-    def _create_nav_button(
-            self, 
-            label: str, 
-            is_checkable: bool = True
-    ) -> NavButton:
-        """Creates and configures a single navigation button.
-
-        The button's icon is determined by its label, which must correspond to an
-        entry in the `SIDEBAR["ICONS"]` configuration.
-
-        Args:
-            label (str): The text label for the button (e.g., "Dashboard").
-            is_checkable (bool): If True, the button is added to an exclusive button group.
-
-        Returns:
-            NavButton: The newly created navigation button.
-        """
-        icon_info = SIDEBAR_ICONS[label.replace(" ", "_").upper()]
-        object_name = f"{label.replace(' ', '_')}Button"
-
-        button = NavButton(
-            file_path = icon_info["PATH"],
-            icon_size = icon_info["SIZE"],
-            variant   = icon_info["DYNAMIC"],
-            text      = label,
-            checkable = is_checkable,
-            height    = 82,
-        )
-        button.setObjectName(object_name)
-        self.mainLayout.addWidget(button)
-
-        if is_checkable:
-            self.nav_button_group.addButton(button)
-
-        return button
-    
-    def _create_avatar(self) -> AvatarLoader:
-        """Creates and adds the avatar and username section to the sidebar."""
-        avatar = AvatarLoader(size=QSize(240, 240))
-        username = QLabel("@username")  # placeholder for username, set dynamically
-        username.setObjectName("UsernameLabel")
-        username.setAlignment(Qt.AlignCenter)
-        avatar_wrapper = create_fixed_wrapper(
-            widgets     = [avatar, username],
-            fixed_width = self.width(),
-            alignment   = Qt.AlignHCenter,
-            direction   = "vertical",
-            margins= (0, 40, 0, 30)
-        )
-        self.mainLayout.addWidget(avatar_wrapper)
